warnings solved
